[py_trace_event] Support clock snapshots in proto trace format
TBMv3-style clock synchronization depends on ClockSnaphost messages
in the trace, that provide links between different time domains.

Telemetry uses BOOTTIME clock domain, but since it sometimes runs
on a different device from Chrome, we can't just use the BOOTTIME
domain from Perfetto. Instead we define a special user clock and provide
a snapshot linking it to the BOOTTIME domain.

This CL adds support for writing ClockSnapshot messages in Perfetto
protobuf format and switches from deprecated TraceEvent timestamps
to TracePacket timestamps. It also uses this new method in Telemetry
to write the clock snapshot in the beginning of the trace.

Bug: chromium:990304
Change-Id: Iaa879f30d9a4e756d7de8efb1f1cdd6ed8e380f7