Use --no-streaming option when installing app on certain devices
This is a reland of crrev.com/c/1990346, plus additional checking for
the adb version in case the streaming options are not available.

The original description:
The current Nexus 9 currently installs apks with default options.
However, it keeps having timeout error when installing big apks,
for example ChromePublicTest.apk (about 150Mb)

This CL will let Nexus 9 install apks with --no-streaming option,
with which apks will be pushed to device and get installed from there.
This hopefully could resolve the timeout issue.

Bug: chromium:1004322
Change-Id: I238f8209cddbc8741c7b3dd095e0e70d1575a88e