Fix bugs where I assume some optional fields were required.
If for instance min in the 'running' message is 0, proto will not
write it out on the wire. Technically the fields in the running
message are optional so I need to treat them as such.

Also throw nicer exceptions if the truly required fields, name and
unit, are not there.

Bug: chromium:1029452
Change-Id: I3fd5b0edc94df636427ec4c5e196dce32cc556b3