Move to protobuf lite.
It didn't work to use protobuf_full in c++ code; WebRTC code uses
protobuf_lite so there's ODR violations happening for many test
test binaries.

I will now check in the generated histogram_pb2.py since it will
be called by webrtc_dashboard_upload.py (and later probably by
process_perf_results in Chromium). It would be better if it was
generated by a GN build rule, but python scripts in WebRTC and
Chromium usually just hack the PYTHONPATH and pull it in, so it's
easier to just check it in.

Bug: chromium:1029452
Change-Id: Idff0f3b636ca6f3ee63ce5475b01ec7456aae232