[Typ] Add ability get string value of an Expectation instance
This CL adds the ability to get the string value of an Expectation
instance. When we get the string value, all asterisks will be escaped
in non glob expectations. For glob expectations, the last asterisk will
not be escaped. Also we will keep the raw strings for the tags and
results from the original string that was parsed for the expectation.
We will also keep the raw tags in the same order they appeared in the
original string. This will prevent modifications to expectations
during WPT imports.

TBR=dpranke@chromium.org
Bug: chromium:986447
Change-Id: I8262107bfe2ec0de2518874af93002dd2b2a63ac