Find multiple points with E-Divisive
We use a heuristic to refine the placement of anomalies, which is to
re-run the E-Divisive change point detection algorithm to catch all the
places where we can find potential changes. This doesn't completely fix
the off-by-one errors, but this means we can find other candidate points
where the changes are also prominent.

For now, we use only one of the potential change points, but we ensure
that we are logging all the change points for the time being. This
allows us to evaluate when debugging alerts by looking for the revisions
in the logs.

R=abennetts@google.com, fancl@chromium.org

reword: Refine Alert Placement
Bug: chromium:1041688
Change-Id: I534399624bd30fe5c9ec81001e1f0f4247ba037a