E-Divisive: Widen Range of Candidates
Previously, we would only consider candidates outside the
min_segment_size margins of the original range. This change allows us to
consider all points in the range except for the first and last elements
as potential culprits in the clustering change detector, and rely on
post-filtering to weed out points that are not meant to be considered.

This also refactors the impementation so that we can have simpler loops
and clearer terminal conditions.

We also add another test-case where we have data that has a misplaced
alert that we can properly place now.

R=abennetts@google.com, fancl@chromium.org

Bug: chromium:1041688
Change-Id: I959f70e8f8f6460c00d570789d5fc00ae455970c