Introduce a repository filter for commits.
This allows us to exclude repositories that Pinpoint shouldn't be
bisecting into, configured through the site configuration system in the
Dashboard.

This is a continuation of https://chromium-review.googlesource.com/c/catapult/+/2018304.
It adds a simple 60s TTL to the in-process caching of the new config value,
moves the config to be a non-namespaced stored object (so that this
unambiguously applies globally, not just to non-internal requests) and fixes
some bugs so that the tests pass.

R=dberris@chromium.org, fancl@chromium.org

Change-Id: I3778cff164094df399bbc4701d7f359dea83a52b
Bug: chromium::1045197