Use auto instead of google::protobuf in the histogram C++ API.
Because there are three incompatible proto implementations downstream,
it's quite hard to get the proto namespace right in this code. We can
use complicated import transforms to fix this as well, but I'd argue
this is a good use of auto since we avoid this problem entirely with
basically no loss of readability.

Bug: chromium:1029452
Change-Id: Ibd3c1bf91e3c5d3eb4c61cba9ab64aac5a4a5e85