Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > Patch Set 1:
> > > 
> > > This is insufficient.
> > > 
> > > We need a contact_email field instead which is required so that we know which team(s) to contact for each subscription.
> > 
> > I'm still not sure adding a contact_email field is an useful idea. As abennetts@ suggested, we can't ensure the email is still in use and it seems won't be better than git blame to see who makes the changes.
> 
> I'm also worried about blocking a reasonable feature request while we figure out the details of how best to maintain a list of current contacts.  If it's quick to add a contact_email field that must be set if notification_email is absent, let's do that, we can always relax restrictions later.
> 
> (If it's not easy to add that field for some reason, then I'd be in favour of just dropping the requirement for an email address immediately so we don't block users while we work out the longer term plans.)

There's a work-around for the required notification_email field as it stands. What we do need is a way to ensure that we can contact folks that are maintaining the configurations if we ever need to make large-scale changes.

Right now notification_email is serving two purposes. We don't have git blame for the current state of the subscriptions so we can't rely on it going forward either.

The solution being proposed in the discussion from which this spawned did involve having a contact_email field that is required, so let's move to that sooner than later. Right now folks have a work-around already (providing a separate alert mailing list from team mailing lists) so if we're going to do this (remove the requirement to have notification_email) we need a way to contact owners now.

Patch-set: 1
Reviewer: Gerrit User 1327821 <1327821@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>