Teaching histograms to deserialize from protos.
For this CL, I thought everything should read the proto as natively
as possible since the proto will be a first-class format (alongside
histogram JSON and the unfinished "serialized" format).

I'm going to generate as many protos as I can from one of our tests
and run through this new code to verify it's correct, but I think
this is a good start at least.

I have to add a protobuf lib to the presubmit python invocations,
and I took the one in third_party/gae_ts_mon/gae_ts_mon/protobuf.
We should consider adding third_party/protobuf to follow the law
of Demeter, but this is OK for now I think. It's just the
presubmit anyway.

Bug: chromium:1029452
Change-Id: Iae3d01fb863e499f2063e4edc85b7f45de8b0eb3