Remove UE bucketing in runtimeStatsTotalMetric
We used to bucket runtimeStatsMetric based on RAIL stages. The idea
behind this was to use the load bucket as a way to measure the loading
time. Though this was always noisy and wasn't used much.

This cl also changes the metric to only loop over threads that could
execute javascript to reduce unnecessary work.

Bug: chromium:958441
Change-Id: I3ecb1a80bcbe12b0697f42b42c91890a83a40850