Devil: throw if app is not installed
This updates GetApplicationDataDirectory() to throw CommandFailedError
if the requested package is not installed. The API doc specifies it may
do so, but the implementation previously would actually just fetch the
launcher app (because this was in the 'dumpsys activity starter' section
of the output).

I also saw an unrelated bug in how this method created a
CommandFailedError (it didn't generate the message string correctly), so
I fixed that as well.

Bug: chromium:1047845
Test: ./devil/bin/run_py_tests
Test: android_webview/tools/record_netlog.py --package not.installed.app
Change-Id: Ia797e430e8305c705eaa67658929cdcdfbcbe4b3