E-Divisive: Reduce permutation testing iterations
This change reduces the maximum iterations from 300 down to 150, cutting
down the CPU time required to compute the probability that there's a
change point in either of the sub-clusters. This is fine now that we are
gathering potential change points and post-filtering instead of picking
a single change.

We also remove some chatty loglines and introduce more indicative ones
that can help in debugging better.

R=abennetts@google.com, fancl@chromium.org

Change-Id: Ic815b94f46053b9c2e646af7af893c355c4be477