Devil: increase timeout when uninstalling WebView
We currently use an increased timeout (300sec) when pushing a new
WebView package onto the device, since WebView is a large app. This uses
the same timeout when pulling the WebView APK off the device, as this is
just about the same amount of work as installing.

This is a speculative fix for timeouts observed in the WebView CTS bots.

Bug: chromium:1050705
Test: None (this is a speculative fix)
Change-Id: I7ca213bb1e1c42b29f86708b36fac3d66d51da15