Enable Webcomponents V0 Polyfill
This CL enables the webcomponent v0 polyfill. Many features are broken
and they will be fixed in later CLs if possible, but the trace is
visible and can be navigated with a mouse. We show a prominent
warning when loaded with the polyfill so people know what is going
on: https://screenshot.googleplex.com/Xv1HOEFjmuY

Tested on local dev server and vulcanized traces.

Bug: chromium:1036492

Change-Id: I90306cc6775d578672069ca23cbeb1086b45575e