Fix getservbyname assumption
Fixes BrowserOptions assuming that socket.getservbyname('ssh') never
fails. While unlikely, it's possible for a system to not have the SSH
service running, likely due to SSH not being installed. In this case,
getservbyname will throw an OSError.

Instead, use a getter that tries to call getservbyname, but logs and
returns an invalid value if it fails.

Bug: chromium:869557
Change-Id: I2fc7c8284cdb249ba59b92709ee48250fa605da4