Fix getservbyname assumption
Fixes BrowserOptions assuming that socket.getservbyname('ssh') never
fails. While unlikely, it's possible for a system to not have the SSH
service running, likely due to SSH not being installed. In this case,
getservbyname will throw an OSError, even if we don't actually need
the SSH port for the tests we're running.

Instead, default --remote-ssh-port to a negative value, and only try
to set it to a proper default if we're actually running a CrOS test
in remote mode.

Bug: chromium:869557
Change-Id: I2fc7c8284cdb249ba59b92709ee48250fa605da4