Regenerate p-value bounds for performance tests
This change regenerates the table of "high" p-values adding tables for
lower comparison magnitudes and higher sample sizes.

-- Additional Background --

Pinpoint uses these high p-value bounds in hypothesis testing for
various sample sizes to determine whether the sample sets we're
comparing show some potential likelihood that there might be a
difference but that we might not have enough samples to say definitively
one way or another.

This is how Pinpoint decides whether there might be a difference between
two sets and that we might need more samples:

  p-value = min(UTest(a, b), KSTest(a, b))
  low_alpha = 0.01  # This is a constant.
  high_alpha = getHighAlpha(scaled_iqr_multiple, min(|a|, |b|))
  if p-value < low_alpha:
    reject null hypothesis
  if p-value < high_alpha:
    re-run experiment with more samples
  else:
    fail to reject null hypothesis

Because the table of values is generated using simulations, we will not
get the same values every time we generate the tables. Because of this
as well, we might end up having "high" threshold values that will
actually be *lower* than the constant alpha. This typically happens
though the more samples you have in the experiment.

R=abennetts@google.com, fancl@chromium.org

Fixed: chromium:1047749
Change-Id: I2545864817531885435411b419ac31c6285b0bb5