[Devil] Use passed in verbosity in logging config.
Currently Devil's devicetests end up with the generic
basicConfig and ignore the --verbose argument that Typ
both advertises and accepts.

Note:
'The functions debug(), info(), warning(), error() and
critical() will call basicConfig() automatically if no
handlers are defined for the root logger.'
from https://docs.python.org/2/library/logging.html#logging.basicConfig

This changelist sets up a basicConfig that respects log level.
This basicConfig also has a reasonable default format s.t.
it is easy to figure out what is going on in a log.

I'm attaching this to issue 383566 simply because that
is the issue I was working on that needed this to help
me understand what was going on.

Bug: chromium:383566
Change-Id: Ia91cc882dda8bb838821ba259719339b316b85d3