WebLayer telemetry: use TelemetryActivity
Using this instead of WebLayerShellActivity means that incidental
changes to the demo app's UI, which can change the size of the viewport,
won't affect website performance in tests.

NOTE: This change should increase the size of the viewport in the perf
tests, so *regressions are expected* as some web pages will load more
resources.

Bug: chromium:1045609
Change-Id: I68a518a5370e502ee0dc8e07bd1e8f4a2acf49ae