[Catapult] Make Deep a top-level owner.
For changelists like this one
https://chromium-review.googlesource.com/c/catapult/+/2050067
it is better if we have more owners so that we can move faster.
Deep is overall TBMv2 and TBMv3 owner and has been maintaining
trace processor and results html, etc. He has view into all the
areas of Catapult and I think he would be a great overall owner.

Note that if you don't feel comfortable reviewing something,
it's always fine to pass it off to another owner!

Bug is unrelated but Catapult requires bugs.

Bug: chromium:1050215
Change-Id: I3460f85e9b29dfb271579e92beb0813c7670cb4f