[trace viewer] Update embedder doc and example
The Go team was following this example and it needed to be updated with
the polyfill information.

Once I touched the file there some presubmit errors, so I changed some
js to make the linter happy, and moved it out of the bin/ directory
because presubmit wants everything there to be executable.

Bug: chromium:1036492
Change-Id: I136ff802803b31177f8bec8424af955d4ff30237