[trace viewer] Add way to hide the polyfill warning
Some embedders wants to show their own UI warning, and hide the warning
inside the trace viewer. Now if window.__hideTraceViewerPolyfillWarning
is true, the UI warning is not shown. Still showing a warning on console
because the fact that polyfill is running is important debugging
information anyone opening the console (and therefore trying to debug
trace viewer) should be aware of.

Bug: chromium:1036492
Change-Id: Ia5b5043386f2a41ef4e820f073fc3ef4ea717c71