[Typ] Set default string arguments to None in Expectation constructor
We need to set the default values of string parameters for the
Expectation constructor to None. Then we need to convert
member variables set to None into empty strings. This will ensure
that string member variables of Expectation instances are equal
to each other whenever the string member variables are not set.

Bug: chromium:986447
Change-Id: I00d7f7f3d135d38f8b30e1b0643f9c4a875d5273