Re-land "WebLayer telemetry: use TelemetryActivity"
This relands commit 515e6dafedc41240ebb9542726e0a2b168ed2957

There's no change from the original. WebLayer's TelemetryActivity had
to be updated to enable remote devtools connections.

here is a pinpoint try job:
https://pinpoint-dot-chromeperf.appspot.com/job/11956eea620000

As expected, some metrics get worse as a result of the increased
viewport size. The console error count increase from 17 to 25,
for example, which is the same increase that was seen after
019a8580aead878f3c912b948, which also increased viewport size by
changing the UI of WebLayerShellActivity. It should be OK to ignore
Pinpoint alerts for this change, as there will be a new effective
baseline going forward.

Bug: chromium:1045609
Change-Id: I952ed1bb1e8f8e333ce3f73b490ec997f079296a