Pinpoint: Preserve parsing specific failures
This change consolidates the logic and tests associated with ensuring
that previous functionality is preserved, and that we're only returning
errors regarding unknown file formats in cases where that's truly the
case. Before this change we had not been preserving semantic errors
raised by the parsing implementations, and now we're able to do that
more specifically.

We're also refactoring this so that we no longer test the previous
specific tests and executions, but rather preserve those for posterity.

We still don't have a simple answer for how we're going to be able to
delete these given that we have Pinpoint jobs that have used these types
in pickled entities in Datastore.

R=abennetts@google.com, fancl@chromium.org

Bug: chromium:1041818
Change-Id: I97517b2134a266e37d140eaf168100f5e8c7bf5e