[typ] Set log level based on --verbose argument.
Also, fix devicetests and catapult_build's typ usages
to pass in sys.argv[1:] so that arguments get parsed properly.

Also, add a log statement to the devicetests base class.
I don't know why this works, but for some reason it
is necessary for getting logging out of the rest of
Devil to happen when you run something like
./run_py_devicetests --test-filter *perf_control_devicetest* -v -v -v -v --passthrough

Background is here
https://chromium-review.googlesource.com/c/catapult/+/2052253
and here
https://chromium-review.googlesource.com/c/catapult/+/1453259

An alternative to this changelist that doesn't modify typ is at
https://chromium-review.googlesource.com/c/catapult/+/2075820

Bug: chromium:1051254, chromium:383566
Change-Id: Idcc539418d6cfefa9ed4f09becc34a5e71a4333f