[Devil] Set up logging level based on --verbose.
This change intentionally doesn't use the runner.context
solution used by
https://chromium-review.googlesource.com/c/catapult/+/2075820
because that feature of typ seems complicated and poorly documented.
Instead, we simply set the logging level before we start. We do this
based on reading the value of the verbose argument by parsing the
arguments ourselves. This means that we are parsing the arguments
twice, but that is actually required anyways because this solution
here https://chromium-review.googlesource.com/c/catapult/+/2052253
doesn't work.

Thanks to Rakib for helping me come up with this much simpler solution!

Bug: chromium:383566
Change-Id: I53f03362e21795caf723256653dfac51a866fe14