Add flag to WPR to disable fuzzy matching during replay.
When creating WPR sessions it is would be nice to sometimes be able to
enforce strict url matching that doesn't fuzzy match if query parameters
aren't exact.

This CL introduces the flag 'exact_match_only' which forces WPR to
not attempt to fuzzy match when a url only has the same path but
inexactly matched query parameters.

It also adds some simple logging additions that will make the level of
match (fuzzy versus exact) more apparent on record and replay.


Bug: chromium:1020982
Change-Id: I5e8317fe783db6d14c51a7f6fd904003a09b7464