Pinpoint: Add logging to results2 generation
This change allows us to look up some information making it into the
results2 html page generation. For now we're just logging the number of
histograms we're seeing, which we can make decisions on how to make this
work better.

We're also adding test coverage for the case where we do have multiple
CLs involved that have associated data, to allow us to see the loglines
show up at least when running tests.

R=abennetts@google.com, fancl@chromium.org

Bug: chromium:1050740
Change-Id: I371eb59cca98f0cc0c0dcd64b2739f86b10ef1ee