[Telemetry] Cleanup unsymbolized minidumps
Updates all the Telemetry test classes that can start browsers to check
for unsymbolized minidumps before and after running a test. Finding
minidumps before is non-fatal, but finding minidumps after is fatal, as
that indicates that some part of the browser crashed and should be
fixed.

The actual checking is currently disabled due to several Chromium tests
needing to be updated to pass with this change, but all the required
code is present.

Bug: chromium:1056235
Change-Id: Ic6165eb0eb8f6bb2c3ca736a63c585d84a7301ec