Fix CrOS minidump pulling on certain devices
Two small fixes for minidump pulling that was affecting certain devices
such as Krane:

1. Forcibly set BREAKPAD_DUMP_LOCATION, as we were relying on Chrome's
   default dump location on CrOS before. Some devices default to
   dumping to /home/chronos/crash, though.
2. Ignore any directories we find in the minidump directory, as trying
   to pull them with GetFile() will fail.

I was unable to confirm that this fully fixes the issues on Krane
devices, as pushing my local Chromium build to the remote device seems
to have caused it to no longer actually start the browser when told to.
However, before pushing that build, it was saving the minidumps to the
correct location, so this *should* work.

Bug: chromium:1055225
Change-Id: If4bdf3400685131ebeab6ef99d0b849b4877ffff