[Telemetry] Enable leftover minidump catching
Enables the functionality for Telemetry and typ-based tests that create
browsers to check for leftover minidumps before and after a test is run.

Catching unsymbolized minidumps beforea test is run, but catching them
after a test is fatal, as that indicates that the test had a non-fatal
crash during the test that should be fixed.

Bug: chromium:1056235
Change-Id: If3eb33a91b9838149e669f8d4951bb885e4122dd