Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> (2 comments)
> 
> > Patch Set 1:
> > 
> > Spent several hours trying to debug why my device wasn't working.. this logging would've helped save a lot of that time.
> 
> Can you elaborate on this a bit?
> 
> I believe that these log lines may have helped you, but I'm concerned that they may be noise more often than not, which I'd like to avoid at a log level that we have on by default.

I had to run a benchmark on an Android which was in an 'unauthorized' state whereas devil expected it to be in a 'ready' state. The benchmark calls out to a benchmark runner, which in turn uses catapult, which in turn uses devil and debugging the mismatch in the device state was very painful through all these layers of plumbing as there were no log messages. `adb devices` listed my device fine which made everything even more confusing. 

Is the expectation that most devices are *not* in the desired state? I'm confused why this logging would happen in the happy path.

Patch-set: 1
Reviewer: Gerrit User 1158954 <1158954@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0