Fix parsing of summary options.
ListFields only returns fields that are set to True. In protos,
false is indistinguishable from not set for basic typesl like boolean.

I tried using wrappers.proto (BoolValue) for the booleans in
summary_options, but there is no plumbing in Chromium that makes it
possible to include that proto. I think it's good enough for now
that if you specify summary options, you have to specify all of them.

Bug: chromium:1029452
Change-Id: I11e06142ca71efb775aa85a3f0f82e63a4d1607f