trace_data: Raise exceptions only after traces are added to artifacts
Currently, when trace data loss occurs, we raise an exception right
after collecting the trace data. This in turns means that we lose the
trace making it hard to debug that data loss.

This patch buffers the exception into the TraceDataBuilder and only
raises it when the builder is cleaned up after traces were added to
the benchmark artifacts.

Bug: chromium:1059767
Change-Id: I2d764f9455b60f6c404c70f30be1d7788b0dadfa