trace_data: Raise exceptions only after traces are added to artifacts
Currently, when any trace data loss occurs, we raise an exception right
after collecting the trace data. This in turns means that we lose all
the trace data, making it hard to debug a specific data loss (e.g. data
loss on a specific thread).

This patch buffers the exception into the TraceDataBuilder and only
raises it when the builder is cleaned up after traces were added to
the benchmark artifacts. Thus, we should retain the trace data.

Bug: chromium:1059767
Change-Id: I2d764f9455b60f6c404c70f30be1d7788b0dadfa