[Telemetry] Don't hard-code temporary CrOS env file
Switches CrOSBrowserFinder to use a random temporary file to store the
existing chrome_dev.conf file while a test is running instead of a
hard-coded path. This fixes issues with devices being left in a bad
state if multiple PossibleBrowsers run SetUpEnvironment.

Also applies this change to the directory where existing minidumps
are stored - no issues with that have been currently hit, but the same
sort of issue could potentially be hit.

Bug: chromium:1059426
Change-Id: I4ee6e559cc09634b804c73c63039834f3045935c