[Telemetry] Fix minidump pulling on L-
Fixes minidump pulling on Android L and before failing due to the fact
that 'stat' was not available on the device. As a workaround,
device_utils.StatPath is used instead in these cases, which only has
resolution down to a minute. In order to prevent pulled minidumps from
erroneously being classified as "too old" because of this resolution
difference, the timestamp is then adjusted by 59 seconds so that it's
as new as possible while still remaining in that one minute window.

Bug: chromium:1056235
Change-Id: Icc4550a84c066695119f1c9e5b7a9efd7ca383e7