Fix dev_server script.
The problem was that PackPoint writes bundled/ into the temp dir AND
adds it as a deployment path. This means that the second temp
deployment dir attemps to symlink tmpdir/bundled to the deployment
path, but tmpdir/bundled is already there.

Also reuse_tmp_dir didn't really work with symlinks. I added code
to refresh the symlinks on the second deployment dir.

I'm going to assume this all works for the normal deploy logic, so
I didn't touch preprocess.py.

Bug: chromium:1053945
Change-Id: I9c0195cdd6eb9f0e2f3c38dbceeb8df1e3b3c42b