[Benchmarking] Limit Pixel2 cpu/gpu frequencies to reduce thermal throttling noise.
If this is identified as a culprit for a performance regression,
please duplicate the all regression bugs to
https://bugs.chromium.org/p/chromium/issues/detail?id=1063432
so that we can track all performance changes caused by this throttling.

This change is expected to reduce noisiness of performance
measurements for these devices.

This may also improve the longevity of these devices by preventing
them from getting too hot.

Bug: chromium:383566
Change-Id: Id09e5bf07a1210a31bf5fc8f895108defb56dbd2