Remove histogram proto JSON converter code.
This code is no longer needed. It was intended as a middle step before
adding full proto support, but it turned out C++ tests could not easily
write out protos as JSON (it requires the
//third_party/protobuf:protobuf_full target, but much WebRTC and
Chromium code depends on light protobufs, so there's an ODR clash).

Also, protos are now fully supported. We just need code to handle
the proto content type in /add_histograms (it assumes JSON at the
moment). This will come in a later CL.

Bug: chromium:1029452
Change-Id: I0cdd2e02ca61d0ed0251b1a618bc3904dc9e7105