Update patch set 2
Patch Set 2:

(16 comments)

Did a more in-depth review.

I really think it's worth breaking this up into a number of changes:

- Change the sheriff-config service for the configuration bits. That's small and we can deploy that without much integration.
- Implement the auto-grouping logic. This requires some cron job configurations too. Let's deliver that as a single change to production.

Then there's a process step, which is to:

- Opt-in some configurations. Let's get one benchmark in the Chromium perf sheriff config.

After that, we have a debugging phase that will allow us to see whether we're getting the alert grouping right. It shouldn't take too long unless we're finding egregious bugs we've left in the implementation. 😊

When that's live:

- Implement the auto-bug-filing feature. This will require some iteration because the details we're putting in the bugs will probably be in flux.
- Support auto-bisect. Work with the initial group of folks that have auto-triage enabled, and then work through auto-bisect configuration.

After all that, we can launch. As post-launch work we need:

- Export of the alert groups to a BigQuery table.
- Dashboard for tracking the effectiveness of Auto-triaged/Bisected bugs.
- Adding support for alert grouping in Telemetry metrics, as additional reserved metric names. This requires coordination with the Telemetry owners (I'm one of them).

I'm hesitant to go landing this as a single change, because I'd really like us to be able to have pieces that we can roll-back and have as little "point of no return" changes.

Patch-set: 2
Reviewer: Gerrit User 1327821 <1327821@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>