Dashboard: Terminate on Permutation Test Convergence
In this change we ensure that if permutation testing suggests new bounds
that are the same as the current bounds, that we terminate further
iterations and log an issue. This fixes the edge-cases we're
encountering in production where permutation testing will keep finding
the same bounds without the algorithm detecting the infinite loop.

R=fancl@chromium.org

Bug: chromium:1067589
Change-Id: I52b6e148658fdf005884f01878be920eea8df23d