Update patch set 3
Patch Set 3:

> Patch Set 3: Code-Review+1
> 
> Lgtm,
> Is there a pinpoint with this change?

There is now: https://pinpoint-dot-chromeperf.appspot.com/job/1747649ee20000 and https://pinpoint-dot-chromeperf.appspot.com/job/128fc00ee20000

From running on the traces for the associated traces in the bug [1], the ratio for the normal case is ~0.38, and for the busy-case, it is ~0.27. So in terms of %change, it's somewhat big, but it would be interesting to see how it varies for the same story in the bots.

[1] https://bugs.chromium.org/p/chromium/issues/detail?id=1052361#c5

Patch-set: 3
Reviewer: Gerrit User 1000111 <1000111@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0