Pinpoint: Do not double-cache Isolate entities
This change ensures that all Isolate lookups happen against the
Datastore, instead of also being cached in memcache. This allows us to
have a single cache (in Datastore) for the mapping of
revisions+builder+target -> isolate that we can manage independently.

We also remove some stale TODOs around the hard-coded mappings for bot
names.

R=abennetts@google.com, fancl@chromium.org

Bug: chromium:1070753
Change-Id: I4eaf8af11050fc918f41b5d0c13541b274ce7b25