NetLogViewer: Add NIK to host cache entries.
DNS entries in net-export logs have included a NIK field since at least
M80. This CL adds code to display that field in the DNS tab. When
partitioning the host cache by NIK is not enabled (currently the
default), NIK is the empty list, and [].toString() is "", so nothing
is displayed in the common case.

NOTRY=true
NOPRESUBMIT=true
Bug: chromium:997049
Change-Id: I07ce9150e58f3b784c938d018229f578e1cac1ca