devil: Add a new option "decrypt" for WaitUntilFullyBooted.
Android devices with full-disk encryption are encrypted on first boot,
and then get decrypted to continue the boot process
(See details in https://bit.ly/3agmjcM).

This can optionally wait for the decryption to complete since it can take
a while for old OSs like Marshmallow.

Note that it will return directly on devices with no encryption or with
file-based encryption.


Bug: chromium:922145, chromium:1057811
Change-Id: I1a4c608a6e439b5b0b0f3b475a2fb77d57c9dedb