Make device_utils.ReadFile faster in the common case.
Simply pulling a file is faster than getting the file's size and then
either cat-ing or pulling.

The cost of getting the size of a file can be a second or more on
remote adb connections (though is typically ~100ms for local
connections).

We should only read by cat when we would otherwise have to copy the file
before pulling it.

Bug: chromium: 1076580
Change-Id: I03e5299f078c0e779b935b9730d6942835ddbc22