Add TestExpectations.is_default_pass
The client may need to know if the test expectation is an explicit
Pass or a default Pass.

Also let [ Slow ] and [ RetryOnFailure ] imply default pass only,
instead of always adding a [ Pass ] expectation. For example, for
[ Linux ] foo/bar [ Failure ]
foo/* [ Slow ]
we should not treat [ Pass ] as a valid expectation on Linux.
Previously we was unable to mark the test always failing on Linux.
(This is not a problem of web test though, because we only allow
[ Slow ] in SlowTests and run_web_tests.py merges the expectation files
by itself.)

Bug: chromium:1072015
Change-Id: I00e1488ca6a0db9acd497289d511e042752d97b5