Add TestExpectations.is_default_pass
This allows the client know if the test expectation is an explicit
Pass or a default Pass.

Also let [ Slow ] and [ RetryOnFailure ] imply default pass only,
instead of always adding a [ Pass ] expectation. For example, for
[ Linux ] foo/bar [ Failure ]
foo/bar [ Slow ]
we should not treat [ Pass ] as a valid expectation on Linux.
Previously we was unable to mark the test always failing on Linux.

Bug: chromium:1072015
Change-Id: I00e1488ca6a0db9acd497289d511e042752d97b5