Pinpoint: Respect extra_test_args from config
This change ensures that we're preserving the extra_test_args configured
per-bot in addition to the extra_test_args provided in the input to
created Pinpoint jobs. This lets us set flags to tests being run based
on the configuration/bot that we're running on. This lets us set
bot-specific flags that will always be present for new Pinpoint jobs.

R=abennetts@google.com, fancl@chromium.org, nuskos@chromium.org

Bug: chromium:1078759
Change-Id: I3feed45410a313a99128a849d3034a817dc0e2ec